Rowid as long
- Adjusted fetching of rowid to be always of type long
-- This also fixes missing duplicate detection

+++ DataTableConverter/Assisstant/ExportHelper.cs
@@ -786,7 +786,7 @@ namespace DataTableConverter
 
             if (count == 0)
             {
-                Dictionary<string, int> pair = DatabaseHelper.GroupCountOfColumn(columnName, tableName);
+                Dictionary<string, long> pair = DatabaseHelper.GroupCountOfColumn(columnName, tableName);
                 DatabaseHelper.DictionaryToTable(pair, columnName, showFromTo, newTable, DatabaseHelper.GetRowCount(tableName));
             }
             else
