+++ DataTableConverter/Classes/CountListboxItem.cs
@@ -5,10 +5,10 @@ namespace DataTableConverter.Classes
 {
     internal class CountListboxItem : IEquatable<CountListboxItem>
     {
-        internal int Count;
+        internal long Count;
         internal object Value;
 
-        public CountListboxItem(int count, object value)
+        public CountListboxItem(long count, object value)
         {
             Count = count;
             Value = value;
